completely change fstore logic again

+++ src/nksrv/lib/fstore/fstore.go
@@ -3,18 +3,13 @@ package fstore
 // abstracts and automates some filestore operations
 
 import (
-	crand "crypto/rand"
-	"encoding/binary"
-	"fmt"
 	"os"
 	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
-	"time"
 
-	"nksrv/lib/pcg"
+	"golang.org/x/xerrors"
 )
 
 type Config struct {
@@ -27,39 +22,12 @@ type FStore struct {
 	privdir string // priv (per instance) folder + path sep if needed
 	privpfx string // priv pfx used for initDirs
 
-	initDirsMu sync.Mutex
-	initDirs   map[string]struct{}
+	// pre-declared dirs
+	declDirsMu sync.RWMutex
+	declDirs   map[string]bool // whether already precreated
 }
 
 // tempfile logic based on stdlib' io/ioutil/tempfile.go
-
-var (
-	rng     pcg.PCG64s
-	rngInit bool
-	rngMu   sync.Mutex
-)
-
-func reseed() {
-	var b [16]byte
-	if _, e := crand.Read(b[:]); e != nil {
-		panic(e.Error())
-	}
-	x, y := binary.BigEndian.Uint64(b[:8]), binary.BigEndian.Uint64(b[8:])
-	x += uint64(os.Getpid())
-	y += uint64(time.Now().UnixNano())
-	rng.Seed(x, y)
-}
-
-func nextSuffix() string {
-	rngMu.Lock()
-	if !rngInit {
-		reseed()
-		rngInit = true
-	}
-	x := rng.Bounded(1e18)
-	rngMu.Unlock()
-	return strconv.FormatUint(1e18+x, 10)[1:]
-}
 
 func badPrivate(p string) bool {
 	return p == "" || (p[0] == '.' && p != ".") ||
@@ -69,14 +37,24 @@ func badPrivate(p string) bool {
 func cleanWSlash(p string) string {
 	p = path.Clean(p)
 	if p != "." {
+		if p == "/" {
+			panic(`plz don't use "/"`)
+		}
 		return p + string(os.PathSeparator)
 	} else {
 		return ""
 	}
 }
 
+const (
+	newMainDirMode = 0777
+	newDirMode     = 0700
+	newFileMode    = 0666
+)
+
 func OpenFStore(cfg Config) (s FStore, err error) {
-	s.initDirs = make(map[string]struct{})
+
+	s.declDirs = make(map[string]bool)
 
 	if cfg.Path == "" || badPrivate(cfg.Private) {
 		panic("incomplete/invalid config")
@@ -93,8 +71,14 @@ func OpenFStore(cfg Config) (s FStore, err error) {
 	s.rootdir = s.privdir[:len(rootdir)]
 	s.privpfx = s.privdir[len(rootdir):]
 
+	if s.rootdir != "" {
+		err = os.MkdirAll(s.rootdir[:len(s.rootdir)-1], newMainDirMode)
+		if err != nil {
+			return
+		}
+	}
 	if s.privdir != "" {
-		err = os.MkdirAll(s.privdir[:len(s.privdir)-1], 0777)
+		err = os.MkdirAll(s.privdir[:len(s.privdir)-1], newDirMode)
 		if err != nil {
 			return
 		}
@@ -108,56 +92,71 @@ func (fs FStore) Main() string {
 	return fs.rootdir
 }
 
-func (fs *FStore) makeDir(root, rpfx, dir string, mode os.FileMode) (err error) {
-	fs.initDirsMu.Lock()
-	defer fs.initDirsMu.Unlock()
+func (fs *FStore) declareDir(root, rpfx, dir string, preserve bool) (err error) {
+	fs.declDirsMu.Lock()
+	defer fs.declDirsMu.Unlock()
 
-	//fmt.Fprintf(os.Stderr, "newdir: %q\n", fs.root+dir)
-	err = os.MkdirAll(root+dir, mode)
-	if err != nil {
-		return
+	// pre-mark
+	fs.declDirs[rpfx+dir] = false
+
+	if preserve {
+		err = os.MkdirAll(root+dir, newDirMode)
+		if err != nil {
+			return
+		}
+
+		fs.declDirs[rpfx+dir] = true
+	} else {
+		err = os.RemoveAll(root + dir)
 	}
-	// mark so if someone removes we'll err
-	fs.initDirs[rpfx+dir] = struct{}{}
 	return
 }
 
-func (fs *FStore) MakeDir(dir string, mode os.FileMode) error {
-	return fs.makeDir(fs.privdir, fs.privpfx, dir, mode)
+func (fs *FStore) DeclareDir(dir string, preserve bool) error {
+	return fs.declareDir(fs.privdir, fs.privpfx, dir, preserve)
 }
 
-func (fs *FStore) MakeGlobalDir(dir string, mode os.FileMode) (err error) {
-	return fs.makeDir(fs.rootdir, "", dir, mode)
+func (fs *FStore) DeclareGlobalDir(dir string, preserve bool) (err error) {
+	return fs.declareDir(fs.rootdir, "", dir, preserve)
 }
 
-func (fs *FStore) removeDir(root, rpfx, dir string) (err error) {
-	fs.initDirsMu.Lock()
-	defer fs.initDirsMu.Unlock()
-
-	err = os.RemoveAll(root + dir)
-	delete(fs.initDirs, rpfx+dir)
-	return
-}
-
-func (fs *FStore) RemoveDir(dir string) error {
-	return fs.removeDir(fs.privdir, fs.privpfx, dir)
-}
-
-func (fs *FStore) RemoveGlobalDir(dir string) error {
-	return fs.removeDir(fs.rootdir, "", dir)
-}
+//var errNotDeclared = errors.New("directory not declared")
 
 func (fs *FStore) ensureDir(fulldir, dir string) (err error) {
-	fs.initDirsMu.Lock()
-	defer fs.initDirsMu.Unlock()
+	// quick read-only peek first
+	fs.declDirsMu.RLock()
+	inited, exists := fs.declDirs[dir]
+	fs.declDirsMu.RUnlock()
 
-	if _, inited := fs.initDirs[dir]; !inited {
-		err = os.MkdirAll(fulldir, 0700)
-		if err != nil {
-			return fmt.Errorf("error at os.MkdirAll: %v", err)
-		}
-		fs.initDirs[dir] = struct{}{}
+	// quick exit conditions determined from peek
+	if inited {
+		// should be already made
+		return
 	}
+	if !exists {
+		// do not allow stuff not in whitelist
+		//return errNotDeclared
+		panic("directory not declared: " + dir)
+	}
+
+	// we'll need to modify it...
+	fs.declDirsMu.Lock()
+	defer fs.declDirsMu.Unlock()
+
+	// maybe it changed since we've looked...
+	inited = fs.declDirs[dir]
+	if inited {
+		return
+	}
+
+	// perform mkdir
+	err = os.MkdirAll(fulldir, newDirMode)
+	if err != nil {
+		return xerrors.Errorf("error at os.MkdirAll: %w", err)
+	}
+
+	// mark as inited
+	fs.declDirs[dir] = true
 
 	return
 }
@@ -166,13 +165,11 @@ func (fs *FStore) makeRndFile(fulldir, pfx, ext string) (f *os.File, err error) 
 	nconflict := 0
 	for i := 0; i < 10000; i++ {
 		name := filepath.Join(fulldir, pfx+nextSuffix()+ext)
-		f, err = os.OpenFile(name, os.O_RDWR|os.O_CREATE|os.O_EXCL, 0666)
+		f, err = os.OpenFile(name, os.O_RDWR|os.O_CREATE|os.O_EXCL, newFileMode)
 		if os.IsExist(err) {
 			nconflict++
 			if nconflict > 10 {
-				rngMu.Lock()
 				reseed()
-				rngMu.Unlock()
 			}
 			continue
 		}
@@ -185,13 +182,11 @@ func (fs *FStore) makeRndDir(fulldir, pfx, ext string) (name string, err error) 
 	nconflict := 0
 	for i := 0; i < 10000; i++ {
 		name = filepath.Join(fulldir, pfx+nextSuffix()+ext)
-		err = os.Mkdir(name, 0700)
+		err = os.Mkdir(name, newDirMode)
 		if os.IsExist(err) {
 			nconflict++
 			if nconflict > 10 {
-				rngMu.Lock()
 				reseed()
-				rngMu.Unlock()
 			}
 			continue
 		}
@@ -219,12 +214,6 @@ func (fs *FStore) NewFile(dir, pfx, ext string) (*os.File, error) {
 
 func (fs *FStore) NewGlobalFile(dir, pfx, ext string) (*os.File, error) {
 	return fs.newFile(fs.rootdir, "", dir, pfx, ext)
-}
-
-const tmpDir = "_tmp"
-
-func (fs *FStore) TempFile(pfx, ext string) (f *os.File, err error) {
-	return fs.NewFile(tmpDir, pfx, ext)
 }
 
 func (fs *FStore) newDir(
